crop_analyzer/crop_analyzer/requirements.txt
Django==5.1.4
Pillow==11.1.0
torch==2.2.1
torchvision==0.17.1
numpy>=1.26.0
scikit-image==0.22.0
